LayoutTests/imported/w3c/web-platform-tests/css/css-fonts/variations/font-weight-interpolation-expected.txt

A
A

FAIL font-weight animation assert_equals: Font weight before animation expected "400" but got "normal"
PASS font-weight transition
